+++ NEWS.rst
@@ -9,6 +9,7 @@ News
 * Handle names imported as aliases like all other used variable names.
 * Rename used_vars to used_names.
 * Use function for determining which imports to ignore.
+* Only try to import each whitelist file once.
 
 
 0.19 (2017-07-20)